Create new_egg_scraper
Filing through all pages of a catalog called Newegg, for specific items based on user input. The pages may be varied but they are set to a fixed value of 10 (on line 13) for quick testing. The upper bound for range in line 13 can be replaced with the comment on line 11 to scrape through all pages of the catalog.